+++ test/system/borrower/repay/creditline_repay.t.sol
@@ -24,9 +24,18 @@ contract CreditlineRepayTest is BaseSystemTest {
         uint256 initialBorrowerBalance = currency.balanceOf(borrower_);
         uint256 initialTrancheBalance = currency.balanceOf(address(reserve));
         uint256 initialCeiling = nftFeed.ceiling(loanId);
-        uint initialNAV = nftFeed.currentNAV();
+        uint256 initialNAV = nftFeed.currentNAV();
         borrower.repay(loanId, amount);
-        assertPostCondition(loanId, tokenId, amount, initialBorrowerBalance, initialTrancheBalance, expectedDebt, initialCeiling, initialNAV);
+        assertPostCondition(
+            loanId,
+            tokenId,
+            amount,
+            initialBorrowerBalance,
+            initialTrancheBalance,
+            expectedDebt,
+            initialCeiling,
+            initialNAV
+        );
     }
 
     function assertPreCondition(uint256 loanId, uint256 tokenId, uint256 repayAmount, uint256 expectedDebt) public {
@@ -44,7 +53,16 @@ contract CreditlineRepayTest is BaseSystemTest {
         assert(currency.balanceOf(borrower_) >= repayAmount);
     }
 
-    function assertPostCondition(uint loanId, uint tokenId, uint repaidAmount, uint initialBorrowerBalance, uint initialTrancheBalance, uint expectedDebt, uint initialCeiling, uint initialNAV) public {
+    function assertPostCondition(
+        uint256 loanId,
+        uint256 tokenId,
+        uint256 repaidAmount,
+        uint256 initialBorrowerBalance,
+        uint256 initialTrancheBalance,
+        uint256 expectedDebt,
+        uint256 initialCeiling,
+        uint256 initialNAV
+    ) public {
         // assert: borrower still loanOwner
         assertEq(title.ownerOf(loanId), borrower_);
         // assert: shelf still nftOwner
@@ -55,29 +73,29 @@ contract CreditlineRepayTest is BaseSystemTest {
             repaidAmount = expectedDebt;
         }
         assert(safeSub(safeSub(initialBorrowerBalance, repaidAmount), currency.balanceOf(borrower_)) <= 1); // (tolerance +/- 1)
-        
+
         // assert: shelf/tranche received funds
         // since we are calling balance inside repay, money is directly transferred to the tranche through shelf
         uint256 newTrancheBalance = safeAdd(initialTrancheBalance, repaidAmount);
         assertEq(currency.balanceOf(address(reserve)), newTrancheBalance, 10); // (tolerance +/- 1)
         // assert: debt amounts reduced by repayAmount (tolerance +/- 1)
-        if (pile.debt(loanId) > safeSub(expectedDebt, repaidAmount)){
+        if (pile.debt(loanId) > safeSub(expectedDebt, repaidAmount)) {
             assert(safeSub(pile.debt(loanId), safeSub(expectedDebt, repaidAmount)) <= 1);
         } else {
             assert(safeSub(safeSub(expectedDebt, repaidAmount), pile.debt(loanId)) <= 1);
         }
         // aseert: creditline
         bytes32 nftID = nftFeed.nftID(loanId);
-        uint maxCeiling = rmul(nftFeed.ceilingRatio(nftFeed.risk(nftID)), nftFeed.nftValues(nftID));
-        uint newCeiling;
+        uint256 maxCeiling = rmul(nftFeed.ceilingRatio(nftFeed.risk(nftID)), nftFeed.nftValues(nftID));
+        uint256 newCeiling;
         if (maxCeiling < pile.debt(loanId)) {
             newCeiling = 0;
         } else {
-             newCeiling = safeSub(maxCeiling, pile.debt(loanId));
+            newCeiling = safeSub(maxCeiling, pile.debt(loanId));
         }
         assertEq(newCeiling, nftFeed.ceiling(loanId));
-        // assert: NAV 
-        assert(safeSub( safeSub(initialNAV, repaidAmount), nftFeed.currentNAV()) <= 1);
+        // assert: NAV
+        assert(safeSub(safeSub(initialNAV, repaidAmount), nftFeed.currentNAV()) <= 1);
     }
 
     function borrowAndRepay(address usr, uint256 nftPrice, uint256 riskGroup, uint256 expectedDebt, uint256 repayAmount)
@@ -95,12 +113,12 @@ contract CreditlineRepayTest is BaseSystemTest {
     }
 
     function testCreditLineCycle() public {
-        uint nftPrice = 200 ether; // -> ceiling 100 ether
-        uint riskGroup = 1; // -> 12% per year
-        uint maxCeiling = 100 ether;
+        uint256 nftPrice = 200 ether; // -> ceiling 100 ether
+        uint256 riskGroup = 1; // -> 12% per year
+        uint256 maxCeiling = 100 ether;
 
         // expected debt after 1 year of compounding
-        (uint loanId, uint tokenId) = createLoanAndWithdraw(borrower_, nftPrice, riskGroup); // borrow max amount
+        (uint256 loanId, uint256 tokenId) = createLoanAndWithdraw(borrower_, nftPrice, riskGroup); // borrow max amount
 
         // supply borrower with additional funds to pay for accrued interest
         topUp(borrower_);
@@ -110,21 +128,21 @@ contract CreditlineRepayTest is BaseSystemTest {
         assertEq(nftFeed.ceiling(loanId), 0); // ceiling = 0 after borrow
         hevm.warp(block.timestamp + 365 days);
         assertEq(nftFeed.ceiling(loanId), 0); // ceiling = 0 after interest accrued
-        uint repaymentAmount = 50 ether;
+        uint256 repaymentAmount = 50 ether;
 
         repay(loanId, tokenId, repaymentAmount, pile.debt(loanId)); // repay half debt
         assertEq(nftFeed.ceiling(loanId), 37999999999999999999); // ceiling 100 - 62 ether
 
-        repay(loanId, tokenId,  pile.debt(loanId), pile.debt(loanId)); // repay rest of debt
+        repay(loanId, tokenId, pile.debt(loanId), pile.debt(loanId)); // repay rest of debt
         assertEq(nftFeed.ceiling(loanId), 100 ether);
-    } 
+    }
 
     function testRepayFullDebt() public {
         uint256 nftPrice = 200 ether; // -> ceiling 100 ether
         uint256 riskGroup = 1; // -> 12% per year
 
         // expected debt after 1 year of compounding
-        uint expectedDebt = 112000000000000000001; // 112 ETH
+        uint256 expectedDebt = 112000000000000000001; // 112 ETH
         uint256 repayAmount = expectedDebt;
         borrowAndRepay(borrower_, nftPrice, riskGroup, expectedDebt, repayAmount);
     }
@@ -134,7 +152,7 @@ contract CreditlineRepayTest is BaseSystemTest {
         uint256 riskGroup = 1; // -> 12% per year
 
         // expected debt after 1 year of compounding
-        uint expectedDebt = 112000000000000000001; // 112 ETH
+        uint256 expectedDebt = 112000000000000000001; // 112 ETH
         // borrower tries to repay twice his debt amount
         uint256 repayAmount = safeMul(expectedDebt, 2);
         borrowAndRepay(borrower_, nftPrice, riskGroup, expectedDebt, repayAmount);
